# docs/openapi-v2.json
{
  "components": {
    "schemas": {
...
      "AccountRequest": {
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "CustomerStateRequest": {
        "properties": {
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "reason"
        ],
        "type": "object"
      },
...
      "Pagination": {
        "properties": {
          "page": {
            "format": "int64",
            "type": "integer"
          },
          "pages": {
            "format": "int64",
            "type": "integer"
          },
          "per_page": {
            "format": "int64",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "page",
          "per_page",
          "pages"
        ],
        "type": "object"
      },
      "PasswordRequest": {
        "properties": {
          "current_password": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "password"
        ],
        "type": "object"
      },
...
      "ReasonRequest": {
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "in": "cookie",
        "name": "apigate_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "HttpOnly browser sessions and CSRF; gateway API keys. Exact decimal USD money, integer usage units.",
    "title": "better-apigate prepaid API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/account": {
      "patch": {
        "operationId": "patch_account",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "properties": {
                  "data": {
                    "properties": {
                      "attributes": {
                        "$ref": "#/components/schemas/AccountRequest"
                      }
                    },
                    "required": [
                      "attributes"
                    ],
                    "type": "object"
                  }
                },
                "required": [
                  "data"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Completed"
          },
          "400": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "401": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "402": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "403": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "404": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "409": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "422": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "429": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "503": {
            "description": "JSON:API error; see errors[0].code and detail"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-admin-required": false
      }
    },
    "/api/v1/account/password": {
      "post": {
        "operationId": "post_account_password",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "properties": {
                  "data": {
                    "properties": {
                      "attributes": {
                        "$ref": "#/components/schemas/PasswordRequest"
                      }
                    },
                    "required": [
                      "attributes"
                    ],
                    "type": "object"
                  }
                },
                "required": [
                  "data"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Completed"
          },
          "400": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "401": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "402": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "403": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "404": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "409": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "422": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "429": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "503": {
            "description": "JSON:API error; see errors[0].code and detail"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-admin-required": false
      }
    },
    "/api/v1/admin/audit": {
      "get": {
        "operationId": "get_admin_audit",
        "parameters": [
          {
            "in": "query",
            "name": "page[number]",
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page[size]",
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "properties": {
                          "attributes": {
                            "properties": {
                              "action": {
                                "type": "string"
                              },
                              "actor_email": {
                                "type": "string"
                              },
                              "actor_id": {
                                "type": "string"
                              },
                              "created_at": {
                                "format": "date-time",
                                "type": "string"
                              },
                              "reason": {
                                "type": "string"
                              },
                              "target_id": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "actor_id",
                              "actor_email",
                              "action",
                              "target_id",
                              "reason",
                              "created_at"
                            ],
                            "type": "object"
                          },
                          "id": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "type",
                          "id",
                          "attributes"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "links": {
                      "additionalProperties": {
                        "type": "string"
                      },
                      "type": "object"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "401": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "402": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "403": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "404": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "409": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "422": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "429": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "503": {
            "description": "JSON:API error; see errors[0].code and detail"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-admin-required": true
      }
    },
    "/api/v1/admin/customers": {
      "get": {
        "operationId": "get_admin_customers",
        "parameters": [
          {
            "in": "query",
            "name": "page[number]",
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page[size]",
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Case-insensitive customer name or email search",
            "in": "query",
            "name": "q",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "properties": {
                          "attributes": {
                            "properties": {
                              "balance": {
                                "description": "Exact USD decimal, stored as integer millionths",
                                "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                                "type": "string"
                              },
                              "created_at": {
                                "format": "date-time",
                                "type": "string"
                              },
                              "email": {
                                "type": "string"
                              },
                              "email_verified": {
                                "type": "boolean"
                              },
                              "frozen": {
                                "type": "boolean"
                              },
                              "name": {
                                "type": "string"
                              },
                              "plan_id": {
                                "type": "string"
                              },
                              "role": {
                                "type": "string"
                              },
                              "shortfall": {
                                "description": "Exact USD decimal, stored as integer millionths",
                                "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                                "type": "string"
                              },
                              "status": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "name",
                              "email",
                              "status",
                              "plan_id",
                              "balance",
                              "shortfall",
                              "frozen",
                              "role",
                              "email_verified",
                              "created_at"
                            ],
                            "type": "object"
                          },
                          "id": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "type",
                          "id",
                          "attributes"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "links": {
                      "additionalProperties": {
                        "type": "string"
                      },
                      "type": "object"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "401": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "402": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "403": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "404": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "409": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "422": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "429": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "503": {
            "description": "JSON:API error; see errors[0].code and detail"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-admin-required": true
      }
    },
    "/api/v1/admin/customers/{id}": {
      "patch": {
        "operationId": "patch_admin_customers_{id}",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "properties": {
                  "data": {
                    "properties": {
                      "attributes": {
                        "$ref": "#/components/schemas/CustomerStateRequest"
                      }
                    },
                    "required": [
                      "attributes"
                    ],
                    "type": "object"
                  }
                },
                "required": [
                  "data"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Completed"
          },
          "400": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "401": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "402": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "403": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "404": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "409": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "422": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "429": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "503": {
            "description": "JSON:API error; see errors[0].code and detail"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-admin-required": true
      }
    },
...
    "/api/v1/admin/customers/{id}/unfreeze": {
      "post": {
        "operationId": "post_admin_customers_{id}_unfreeze",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "properties": {
                  "data": {
                    "properties": {
                      "attributes": {
                        "$ref": "#/components/schemas/ReasonRequest"
                      }
                    },
                    "required": [
                      "attributes"
                    ],
                    "type": "object"
                  }
                },
                "required": [
                  "data"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Completed"
          },
          "400": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "401": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "402": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "403": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "404": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "409": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "422": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "429": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "503": {
            "description": "JSON:API error; see errors[0].code and detail"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-admin-required": true
      }
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "post_auth_logout",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Completed"
          },
          "400": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "401": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "402": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "403": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "404": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "409": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "422": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "429": {
            "description": "JSON:API error; see errors[0].code and detail"
          },
          "503": {
            "description": "JSON:API error; see errors[0].code and detail"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-admin-required": false
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}